+++ setup.py
@@ -50,7 +50,7 @@ setup(
         "Tracker": "https://github.com/passren/PyDynamoDB/issues",
     },
     packages=find_packages(include=[
-        "pydynamodb", "pydynamodb.sqlalchemy_dynamodb", "pydynamodb.sql", "pydynamodb.superset"
+        "pydynamodb", "pydynamodb.sql", "pydynamodb.sqlalchemy_dynamodb", "pydynamodb.superset_dynamodb"
     ]),
     include_package_data=True,
     install_requires=install_requires,
