+++ tests/conftest.py
@@ -172,7 +172,7 @@ def dict_cursor(request):
 
 @pytest.fixture
 def superset_cursor(request):
-    from pydynamodb.superset.pydnamodb import SupersetCursor
+    from pydynamodb.superset_dynamodb.pydnamodb import SupersetCursor
 
     yield from _cursor(SupersetCursor, request)
